pyproject.toml
[project]
name = "matrix-e2ee-bridge"
version = "0.1.0"
description = "E2EE Matrix bridge CLI for Claude Code ↔ OpenClaw communication"
requires-python = ">=3.10"
license = "GPL-3.0-or-later"
readme = "README.md"
dependencies = [
    "matrix-nio[e2ee]>=0.24.0",
    "python-olm>=3.2.0",
    "cachetools>=5.0.0",
    "peewee>=3.0.0",
    "atomicwrites>=1.4.0",
]

[project.scripts]
matrix-bridge = "matrix_bridge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/matrix_bridge"]
